% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSESA.R
\name{FindSerotype}
\alias{FindSerotype}
\title{Find the serotype based on the analysis of the new spacers.}
\usage{
FindSerotype(csesa1 = NA, csesa2 = NA)
}
\arguments{
\item{csesa1}{The new spacer of the first sequence.}

\item{csesa2}{The new spacer of the second sequence.}
}
\value{
The data frame which represents the serotype.
}
\description{
Find the serotype based on the analysis of the new spacers.
}
