% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSESA.R
\name{GetAllNewSpacers}
\alias{GetAllNewSpacers}
\title{Get the new spacers from the molecular sequence and its reverse complement.}
\usage{
GetAllNewSpacers(molecular.seq = NULL)
}
\arguments{
\item{molecular.seq}{The molecular sequence.}
}
\value{
The vector of the new spacers, which is extracted from the molecular sequence and its reverse complement.
}
\description{
Get the new spacers from the molecular sequence and its reverse complement.
}
\note{
If there doesn't exist any new spacer, the function would return NA.
}
