% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSESA.R
\name{GetNewSpacer}
\alias{GetNewSpacer}
\title{Get the new spacer from the molecular sequence.}
\usage{
GetNewSpacer(molecular.seq = NULL)
}
\arguments{
\item{molecular.seq}{The molecular sequence.}
}
\value{
The new spacer sequence as a string.
}
\description{
Get the new spacer from the molecular sequence.
}
\examples{
GetNewSpacer("AGAGGCGGACCGAAAAACCGTTTTCAGCCAACGTAT")

}
