% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSESA.R
\name{GetNewSpacerCode}
\alias{GetNewSpacerCode}
\title{Get the new spacer from the molecular sequence and map it to the code.}
\usage{
GetNewSpacerCode(molecular.seq = NULL)
}
\arguments{
\item{molecular.seq}{The molecular sequence.}
}
\value{
The new spacer code as a string.
}
\description{
Get the new spacer from the molecular sequence and map it to the code.
}
