% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSESA.R
\name{GetReverseComplement}
\alias{GetReverseComplement}
\title{Return the reverse complement of the sequence.}
\usage{
GetReverseComplement(x)
}
\arguments{
\item{x}{The input sequence.}
}
\value{
The reverse complement sequence as a string.
}
\description{
Return the reverse complement of the sequence.
}
