% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSESA.R
\name{GetStr}
\alias{GetStr}
\title{Get the information string from the CSESA s3 object.}
\usage{
GetStr(csesa)
}
\arguments{
\item{csesa}{The S3 object CSESA.}
}
\value{
The string record the newly spacers and serotype information.
}
\description{
Get the information string from the CSESA s3 object.
}
