% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSESA.R
\name{PCR}
\alias{PCR}
\title{Get the CSESA obeject through the two sequence.}
\usage{
PCR(seq1, seq2, out.file)
}
\arguments{
\item{seq1}{The first DNA sequence.}

\item{seq2}{The second DNA sequence.}

\item{out.file}{Into which results will be saved if this value is set. Otherwise results will be displayed on the screen.}
}
\description{
Get the CSESA obeject through the two sequence.
}
