% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSESA.R
\name{ReadInFile}
\alias{ReadInFile}
\title{Read the three types of input file.}
\usage{
ReadInFile(file.name)
}
\arguments{
\item{file.name}{The input file name.}
}
\value{
The molecular sequence as a string.
}
\description{
Read the three types of input file.
}
