% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CSESA.R
\name{WGS}
\alias{WGS}
\title{Find the serotype based on the analysis of the new spacers.}
\usage{
WGS(file)
}
\arguments{
\item{file}{The input fasta file.}
}
\value{
The two DNA molecular sequence.
}
\description{
Find the serotype based on the analysis of the new spacers.
}
